优化一维存储器窗口的操作逻辑

+++ Gigavolt.Expand/MoreMemoryBanks/ListMemory/EditGVListMemoryBankDialog.cs
@@ -12,6 +12,7 @@ namespace Game {
 
         public TextBoxWidget m_linearTextBox;
         public TextBoxWidget m_colCountTextBox;
+        public TextBoxWidget m_offsetTextBox;
         public TextBoxWidget m_widthTextBox;
         public TextBoxWidget m_heightTextBox;
         public LabelWidget m_colCountTextLabel;
@@ -32,6 +33,7 @@ namespace Game {
             m_linearTextBox = Children.Find<TextBoxWidget>("EditGVListMemoryBankDialog.LinearText");
             m_colCountTextBox = Children.Find<TextBoxWidget>("EditGVListMemoryBankDialog.ColCount");
             m_colCountTextLabel = Children.Find<LabelWidget>("EditGVListMemoryBankDialog.ColCountLabel");
+            m_offsetTextBox = Children.Find<TextBoxWidget>("EditGVListMemoryBankDialog.Offset");
             m_widthTextBox = Children.Find<TextBoxWidget>("EditGVListMemoryBankDialog.Width");
             m_heightTextBox = Children.Find<TextBoxWidget>("EditGVListMemoryBankDialog.Height");
             m_IDLabel = Children.Find<LabelWidget>("EditGVListMemoryBankDialog.ID");
@@ -58,6 +60,7 @@ namespace Game {
                 }
                 m_widthTextBox.Text = m_memoryBankData.m_width.ToString();
                 m_heightTextBox.Text = m_memoryBankData.m_height.ToString();
+                m_offsetTextBox.Text = m_memoryBankData.m_offset.ToString();
             }
         }
 
@@ -65,54 +68,61 @@ namespace Game {
             if (m_okButton.IsClicked) {
                 if (uint.TryParse(m_widthTextBox.Text, out uint width)
                     && uint.TryParse(m_heightTextBox.Text, out uint height)
+                    && uint.TryParse(m_offsetTextBox.Text, out uint offset)
                     && int.TryParse(m_colCountTextBox.Text, out int length)
                     && length > 0) {
-                    m_memoryBankData.m_width = width;
-                    m_memoryBankData.m_height = height;
+                    if (m_enterString != m_linearTextBox.Text) {
+                        try {
+                            m_memoryBankData.String2Data(m_linearTextBox.Text, length);
+                            m_memoryBankData.m_width = width;
+                            m_memoryBankData.m_height = height;
+                            m_memoryBankData.m_offset = offset;
+                            m_memoryBankData.SaveString();
+                            Dismiss(true);
+                        }
+                        catch (Exception ex) {
+                            string error = ex.ToString();
+                            Log.Error(error);
+                            DialogsManager.ShowDialog(
+                                null,
+                                new MessageDialog(
+                                    LanguageControl.Error,
+                                    LanguageControl.Get(GetType().Name, 2),
+                                    "OK",
+                                    null,
+                                    null
+                                )
+                            );
+                        }
+                    }
+                    else {
+                        m_memoryBankData.m_width = width;
+                        m_memoryBankData.m_height = height;
+                        m_memoryBankData.m_offset = offset;
+                        if (m_memoryBankData.Data.Count > length) {
+                            m_memoryBankData.Data.RemoveRange(length, m_memoryBankData.Data.Count - length);
+                            m_memoryBankData.m_updateTime = DateTime.Now;
+                            m_memoryBankData.m_dataChanged = true;
+                            m_memoryBankData.SaveString();
+                            Dismiss(true);
+                        }
+                        else {
+                            Dismiss(false);
+                        }
+                    }
                 }
                 else {
                     DialogsManager.ShowDialog(
                         null,
                         new MessageDialog(
                             LanguageControl.Error,
-                            LanguageControl.Get(GetType().Name, 2),
+                            LanguageControl.Get(GetType().Name, 3),
                             "OK",
                             null,
                             null
                         )
                     );
                     return;
-                }
-                if (m_enterString != m_linearTextBox.Text) {
-                    try {
-                        m_memoryBankData.String2Data(m_linearTextBox.Text, length);
-                        m_memoryBankData.SaveString();
-                        Dismiss(true);
-                    }
-                    catch (Exception ex) {
-                        string error = ex.ToString();
-                        Log.Error(error);
-                        DialogsManager.ShowDialog(
-                            null,
-                            new MessageDialog(
-                                LanguageControl.Error,
-                                LanguageControl.Get(GetType().Name, 2),
-                                "OK",
-                                null,
-                                null
-                            )
-                        );
-                    }
-                }
-                else {
-                    if (m_memoryBankData.Data.Count > length) {
-                        m_memoryBankData.Data.RemoveRange(length, m_memoryBankData.Data.Count - length);
-                        m_memoryBankData.m_updateTime = DateTime.Now;
-                        m_memoryBankData.m_dataChanged = true;
-                    }
-                    else {
-                        Dismiss(false);
-                    }
                 }
             }
             if (m_copyIDButton.IsClicked) {
